requirements.txt
delta-lake-reader[azure]
pyarrowfs-adlgen2
adlfs
mlplatformutils
fastapi[all]
python-dotenv